pyproject.toml
[project]
name = "fastapi-sync-async"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "opentelemetry-api>=1.36.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.36.0",
    "opentelemetry-instrumentation-fastapi>=0.57b0",
    "opentelemetry-sdk>=1.36.0",
    "prometheus-client>=0.22.1",
    "uvicorn>=0.35.0",
]
